Testing all models for forward
