+++ conf/models/registration/pointnet2.yaml
@@ -8,7 +8,7 @@ models:
       radii: [[0.2], [0.4]]
       nsamples: [[64], [64]]
       down_conv_nn:
-        [[[FEAT, 32, 32, 64]], [[64+3, 64, 128, 128]]]
+        [[[FEAT +3, 32, 32, 64]], [[64+3, 64, 128, 128]]]
     mlp_cls:
       nn: [128, 128, 32]
       dropout: 0.5
