+++ conf/models/segmentation/pointnet.yaml
@@ -1,7 +1,7 @@
 models:
     PointNet:
         class: pointnet.PointNet
-        conv_type: "DENSE"
+        conv_type: "MESSAGE_PASSING"
         input_stn:
             local_nn: [3, 64, 128, 1024]
             global_nn: [1024, 512, 256]
