+++ conf/models/segmentation/rsconv.yaml
@@ -99,7 +99,7 @@ models:
                     [512 * 3 + 256 * 3, 512, 512],
                     [128 * 3 + 512, 512, 512],
                     [64 * 3 + 512, 256, 256],
-                    [256 + 3 + 3, 128, 128],
+                    [256 + FEAT , 128, 128],
                     [],
                 ]
             skip: True
