+++ docs/conf.py
@@ -28,10 +28,30 @@ author = "Thomas Chaton and Nicolas Chaulet"
 # Add any Sphinx extension module names here, as strings. They can be
 # extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
 # ones.
-extensions = ["sphinx_rtd_theme", "sphinx.ext.autosectionlabel", "sphinx.ext.autodoc" ,'sphinx.ext.mathjax',    'sphinx.ext.githubpages',
-    'sphinx.ext.napoleon']
+extensions = [
+    "sphinx_rtd_theme",
+    "sphinx.ext.autosectionlabel",
+    "sphinx.ext.autodoc",
+    "sphinx.ext.mathjax",
+    "sphinx.ext.githubpages",
+    "sphinx.ext.viewcode",
+    "sphinx.ext.napoleon",
+]
 autosectionlabel_prefix_document = True
-autodoc_mock_imports = ["torch_scatter", "torch_sparse", "torch_cluster", "torch_points", "torch", "torch_geometric", "sklearn", "omegaconf","etw-pytorch-utils","tqdm", "hydra", "matplotlib"]
+autodoc_mock_imports = [
+    "torch_scatter",
+    "torch_sparse",
+    "torch_cluster",
+    "torch_points",
+    "torch",
+    "torch_geometric",
+    "sklearn",
+    "omegaconf",
+    "etw-pytorch-utils",
+    "tqdm",
+    "hydra",
+    "matplotlib",
+]
 
 # Add any paths that contain templates here, relative to this directory.
 templates_path = ["_templates"]
@@ -60,4 +80,3 @@ html_theme_options = {
 # relative to this directory. They are copied after the builtin static files,
 # so a file named "default.css" will overwrite the builtin "default.css".
 html_static_path = ["_static"]
-
