+++ docs/index.rst
@@ -3,7 +3,7 @@
    You can adapt this file completely to your liking, but it should at least
    contain the root `toctree` directive.
 
-:github_url: "https://github.com/nicolas-chaulet/deeppointcloud-benchmarks"
+:github_url: https://github.com/nicolas-chaulet/deeppointcloud-benchmarks
 
 Deep Point Cloud Benchmark
 ======================================================
