+++ src/models/registration/pointnet2.py
@@ -91,4 +91,5 @@ class PatchPointNet2_D(BackboneBasedModel):
         """Calculate losses, gradients, and update network weights; called in every training iteration"""
         # caculate the intermediate results if necessary; here self.output has been computed during function <forward>
         # calculate loss given the input and intermediate results
-        self.loss.backward()  # calculate gradients of network G w.r.t. loss_G
+        if hasattr(self, "loss") and self.loss:
+            self.loss.backward()  # calculate gradients of network G w.r.t. loss_G
