+++ src/models/segmentation/pointnet.py
@@ -3,7 +3,7 @@ import logging
 
 from src.core.base_conv.base_conv import *
 from src.core.common_modules.base_modules import *
-
+from src.utils.config import ConvolutionFormatFactory
 from src.modules.PointNet import *
 from src.models.base_model import BaseModel
 from src.utils.model_building_utils.resolver_utils import flatten_dict
@@ -14,19 +14,15 @@ log = logging.getLogger(__name__)
 class PointNet(BaseModel):
     def __init__(self, opt, type, dataset, modules_lib):
         super().__init__(opt)
-
-        self.has_fixed_points_transform = (
-            dataset.has_fixed_points_transform if hasattr(dataset, "has_fixed_points_transform") else False
-        )
         self.pointnet_seg = PointNetSeg(**flatten_dict(opt))
+        self._is_dense = ConvolutionFormatFactory.check_is_dense_format(self.conv_type)
 
     def set_input(self, data, device):
         data = data.to(device)
         self.input = data
         self.labels = data.y
 
-        batch_size = len(data.__slices__["pos"]) - 1
-        self.pointnet_seg.set_scatter_pooling(not self.has_fixed_points_transform, batch_size)
+        self.pointnet_seg.set_scatter_pooling(not self._is_dense)
 
     def forward(self):
